refactor: move db access methods to util/helper.js
